Preliminary target RWX permissions checking on loads and stores, quick and dirty RIGEL_CFG_NUM=2 Heartbeat.
The backing store now holds lists of target address intervals that
are readable, writable, and executable.  It checks all target read
and write accesses against these lists and throws an exception if
the access is disallowed.
FIXME should this checking be moved to a separate class?

These lists of intervals are populated by the ELF loader.  Executable
ELF sections are marked as executable, static data sections are marked
as read-only or read-write according to the ELF section header flags,
and the heap and stack areas are marked read-write.  We mark
3*ISSUE_WIDTH words past the end of each executable ELF section as
executable, to prevent spurious exceptions when we speculatively
fetch past the end of the program before we resolve the branch.

We detect the end of the static data sections and the beginning of
the heap the same way target code does; we have a special symbol
_end that the linker script places there for us, and we find its
address by iterating over the ELF symbol table.

NOTE: Instead of read_word() and write_word(), backing_store now
exposes read_instr_word(), read_data_word(), read_host_word(),
and write_word().  read_host_word() is used for reads that may be
unsafe for the target to do, but we want to do for simulator-specific
purposes (like dumping memory, or printing an address for the user
in interactive mode).  The other three are all checked against the
valid intervals.  If any of the 3 sets of intervals are empty,
any access of that type will succeed.

FIXME Split write_word() into write_data_word() and write_host_word()
so that we can do cheating writes for e.g. putting syscall results
back in memory once we have more restrictive permissions on stack
and heap.