Rename a few memory-related files, make MemoryTiming* Components.
Also made the RigelSim object a Component, because *it* should be
the root of the component hierarchy, not the Chip, because the
MemoryTiming object is a peer to Chip, not a subordinate or parent.

In time, this assumption may change and MemoryTiming may become
a child of Chip, but whatever object contains the backing store
will still need to be a peer to Chip and a child of RigelSim.

MemoryTiming* don't implement most of the ComponentBase() methods
yet, but they can at least be clocked, heartbeated, etc. within
the common framework.