Replace map with unordered_map to store precoded static instr info.
maps are backed by rbtrees and are slower than hashtable-backed maps
when there are no ordering requirements.  this patch speeds up
most simulations by 15-30%.

right now, i statically resize the hash table to have as many buckets
as static instructions in the binary; this heuristic may need to be
tuned, or we may want to do without it altogether to prevent coupling
between the ELF loader and PipePacket stuff.